Extract legend code
